chore(suite-native): split ConnectPopupScreen
